Extract code to Game.js
